search-agent/pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "langchain-tavily>=0.2.11",
    "pylint>=3.3.8",
    "python-dotenv>=1.1.1",
]
[tool.pylint.MASTER]
ignore = [
    "migrations",
    ".venv",
    "build",
    "generated_code",
    "legacy",
    "__pycache__",
]
